some style fixes
