+++ assignment1/isPermutation.py
@@ -2,7 +2,7 @@ def isPermutation(str1, str2):
     """Function that checks if one string is a permutation of the other.
 
     Args:
-        str1: fist string;
+        str1: first string;
         str2: second string.
 
     Return:
@@ -15,15 +15,15 @@ def isPermutation(str1, str2):
     str2 = str2.lower()
     if str1 == str2:
         return True
-    sybsFromStr1 = dict()
+    symbolsFromStr1 = dict()
     for symbol in str1:
-        sybsFromStr1[symbol] = sybsFromStr1.get(symbol, 0) + 1
+        symbolsFromStr1[symbol] = symbolsFromStr1.get(symbol, 0) + 1
     for symbol in str2:
-        if symbol not in sybsFromStr1.keys():
+        if symbol not in symbolsFromStr1:
             return False
-        sybsFromStr1[symbol] -= 1
-    for symbol in sybsFromStr1:
-        if sybsFromStr1[symbol]:
+        symbolsFromStr1[symbol] -= 1
+    for symbol in symbolsFromStr1:
+        if symbolsFromStr1[symbol]:
             return False
     return True
